Decoder chaining!
I've wanted to do this for a while and thought it would be a bugger job than it actually was, so I'm thrilled it turned out to be so easy.

Chain decoders changes the target container of each decoder button to be the Text decoder textarea.
So if you have text -> hex -> base64 you would paste the base64 into the textarea, activate the chain, then press the relevant decode button to quickly solve the puzzle.
Basically, it saves time copy/pasting