Liquid headers
Moved the file headers out of JavaScript and into a JSON file that can be parsed by liquid, then back into JS again so we can use the file for the actual checks.

Added a few extra headers and set up better labels and file extensions.

Updated the styling of the success / failure messages