Color generator
Generate blocks of color based on any string of information. Data is converted to hex, which is then chunked up into groups of 6 and present as styled blocks. A dropdown is present to pad out any remaining groups that don't add up to 6 characters. It will either pad with a 0 or an F