Clean up of tags/labels
Instead of a hard coded list of tags, set up a global data object of tags (labels, to prevent front matter overlap in naming). This means we can add extra info to tags and also use this to build a search filter.

Added some functionality to ensure that any tags added to front matter exist in the labels.json file, so throw an error before build.

Additionally, added a cleanup function to clear out the contents of _site/ each build, so that it doesn't get filled up with old, unused, or outdated files

Removed changelog as a front matter tag, as we're no longer actively using that as the collection filter