Another milestone update
We've got the landing pages dynamically loading in content from the generated collections - dynamically generating landing pages failed for now, so went back to the manual route of individual pages, though not at root level.

To address the issue of duplicate headings when you go to an individual tool page, I got sneaky and set up a filter. Start off with the headings hidden so they don't show on the dedicated page. When loading them into the landing page, run the filter to remove the hidden class

Fixed 404 references due to structure naming change
Changed targeting for section toggles for restoring localstorage options. As the toggles are hidden on each direct tool page, disable the toggle here by not including those elements