Navbar menu on mobile
A new navigation icon design for mobile layouts has been added for the 4 icons in the nav bar. Due to limited space on mobile, there may be some instances where the icons wrap to the next line (even after removing the tag line). This change moved the icons into a new dropdown menu. As sub-menus are not currently supported within Bootstrap, the theme options have been exposed at the bottom but work in exactly the same way.