Enhanced theme storage
Made an adjustment to how themes are stored, moving to a JSON object instead of just a plain text string. This means that things like descriptions can be stored and displayed. An example is on the Privacy policy, where the theme will be listed with its friendly name and description.

Additionally, handled the migration of old theme values to new ones.

Updated privacy with some checks to see if the returned data is a JSON object, if so parse accordingly