Improved handling of themes and migration of settings
Moved theme file to a module so it can make use of imported tools
Scrapped legacy theme migration, in favor of just clearing out all legacy storage items and starting fresh - this will be removed in a future update

Changed default fallback theme to "auto" instead of light, as a better respect of user preferences
Better handling of default delay in showToast
Added autohide parameter customization to showToast
Removed has-table class from the local storage table on the privacy page. The class was set on the hash table to fix text wrapping. which is not necessary here